+++ src/app/utils/dataCovertor/convertToCesiumObject.ts
