+++ src/app/utils/websocket-supplier/websocket-supplier.ts
@@ -1,7 +1,4 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class socketIoProviderService {
+export class WebSocketSupplier {
   private _socket = io.connect('http://localhost:3000');
 
   get(): any {
